feat: move inventory_gen_steps.py
