+++ cucumber_tests/shared_steps/inventory_gen_steps.py
