+++ cucumber_tests/step_defs/test_environment_inventory_generation.py
@@ -1,5 +1,5 @@
 ﻿from pytest_bdd import scenarios
-from cucumber_tests.step_defs.inventory_gen_steps import *
+from cucumber_tests.shared_steps.inventory_gen_steps import *
 from cucumber_tests.shared_steps.common_steps import *
 from cucumber_tests.shared_steps.unified_pipeline_steps import *
